git_pull.pyとreload.pyをadmin.py(class Admin)にマージ
